requirements.txt
fastapi
uvicorn
python-multipart
pymupdf==1.25.5
pytesseract==0.3.13
pdf2image==1.17.0
Pillow
google-genai==1.9.0
python-dotenv